Added AoE tower and projectile
AoE tower can now shoot projectiles that cause damage within a set
radius of the target. The projectile needs to directly hit an enemy, and
if it does, then it will also hit other enemies around it within the set
radius. There is no animation for this effect yet but you can visually
see that multiple enemies are being damaged (such as the blue ghosts).